Add an observable JSON object and widget
