+++ src/common/observablejson.css
@@ -6,7 +6,7 @@
 |----------------------------------------------------------------------------*/
 
 
-.jp-MetadataEditor {
+.jp-ObservableJSONWidget {
   padding-left: 10px;
   padding-right: 10px;
   display: flex;
@@ -15,7 +15,7 @@
 }
 
 
-.jp-MetadataEditor-host {
+.jp-ObservableJSONWidget-host {
   flex: 1 1 auto;
   border: var(--jp-border-width) solid #cfcfcf;
   border-radius: 0px;
@@ -24,13 +24,13 @@
 }
 
 
-.jp-MetadataEditor.jp-mod-error .jp-MetadataEditor-host {
+.jp-ObservableJSONWidget.jp-mod-error .jp-ObservableJSONWidget-host {
   border-color: red;
   outline-color: red;
 }
 
 
-.jp-MetadataEditor-buttons {
+.jp-ObservableJSONWidget-buttons {
   font-family: FontAwesome;
   flex: 1 0 auto;
   min-height: 13px;
@@ -38,14 +38,14 @@
 }
 
 
-.jp-MetadataEditor-commitButton::before {
+.jp-ObservableJSONWidget-commitButton::before {
   content: '\f0c7'; /* save */
   float: right;
   margin-right: 20px;
 }
 
 
-.jp-MetadataEditor-revertButton::before {
+.jp-ObservableJSONWidget-revertButton::before {
   content: '\f0e2'; /* undo */
   float: right;
   margin-right: 4px;
